python/pyproject.toml
[project]
name = "lpic-langserver"
version = "0.0.1"
description = "A Literate Programming in ConTeXt langauge server"
authors = [
    {name = "Stephen Gaito", email = "stephen@perceptisys.co.uk"},
]
dependencies = [
    "pyyaml>=6.0",
    "pkg @ git+https://github.com/litProgConTeXt/conTeXt-langServer.git@main#subdirectory=python",
]
requires-python = ">=3.10"
license = {text = "Apache2.0"}

[project.scripts]
lpic = "lpicLangServer.lpicCli.cli:cli"
lpic-langserver = "lpicLangServer:langserver.cli"

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "pytest>=7.2.1",
    "pytest-asyncio>=0.20.3",
    "aiofiles>=23.1.0",
    "pytest-mock>=3.10.0",
    "deepdiff>=6.2.3",
]
